pyproject.toml
[tool.poetry]
name = "event-service"
version = "0.1.0"
description = ""
authors = ["Danil Emurashin"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
yoyo-migrations = "^8.2.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-settings = "^2.1.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.16"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
ruff = "^0.1.11"
loguru = "^0.7.2"
grpcio = "^1.60.0"
grpcio-tools = "^1.60.0"
python-multipart = "^0.0.6"
boto3 = "^1.34.30"
prometheus-client = "^0.19.0"
py-grpc-prometheus = "^0.7.0"
grpcio-health-checking = "^1.60.1"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
